proposal: individual membership to stage 1
This refactors the current README.me to break out the bits
that are measuring the proposal VS. stuff that could be
turned into a charter or required documents to outline the program.

If we have consensus to move forward with this we should land as
stage 1 and iterate before going for stage 2.