proposal: expectations to stage 2 (#92)
I believe that the major bits we didn't have consensus on have been
resolved in other proposals. This PR moves expectations to Stage 2,
so we can hand it off to the boards and hopefully move it to Stage 3
in a timely fashion.

Very minor tweaks from Stage 1, specifically removing language about
bits that we didn't currently have consensus on.