doc: incubating projects must have a champion
We had originally had "identifying a champion" as a section in the
application but removed it to make the process more inclusive.

This change re-adds the champion as an entrance criteria for the
incubation phase. Without a champion in incubation it will be quite
a bit harder for projects to be successful. Further, if no member
of the CPC is willing to sponsor / champion a project that would
be a fairly clear sign that the project should not enter incubation.