Update contact email for new project proposals
This commit changes the email address to which new projects proposals are sent.

Currently, new project proposals are sent directly to the mailing list where
they are also discussed.  By having the proposal first come to a person, it
provides a first-round filter to ensure all information in the application is
complete, and also ensures a clean separation between the project applicant and
the committee which is discussion the application.  The OpenJS Foundation
program manager (currently me) would then be responsible for forwarding
applications on to the (private) discussion list.

Signed-off-by: Brian Warner <brian@bdwarner.com>