Add Foundation-wide copyright guidance
This commit adds Foundation-wide copyright guidance on how to add copyright
notices to source code headers.  This document is based on the guidance document
from CNCF, available here:
https://github.com/cncf/foundation/blob/master/copyright-notices.md

To ensure clarity, do *NOT* modify an existing copyright line unless you put it
there, or have explicit permission from the one who did.  This guidance only
applies to copyright notices which are *APPENDED* to the existing list of
notices.

In addition, the OpenJS Foundation Board of Directors recently approved
standardized website footers for projects to use.  These include the copyright
notice, as well as trademark notices and links to Foundation-wide policies.  One
notice is for projects related to Node.js, and includes a line referencing the
trademark license.  The other is for projects which are unrelated to Node.js,
and omits the line.

If you have any questions, please reach out to legal-questions@openjsf.org

Signed-off-by: Brian Warner <brian@bdwarner.com>