Add IP Policy guidance
The goal of this IP Policy guidance is to make it really simple for
projects to comply with the Foundation's IP Policy.

This document is non-exhaustive by design. It is meant to address
the most common scenarios, not every corner-case.

It is also opinionated: it favors community norms and best practices
in areas where the Foundation's IP Policy provides leeway to do so.

This document is based on prior work from CNCF, available here:
https://github.com/cncf/foundation/blob/master/copyright-notices.md,
an earlier pull request by @brianwarner, available here:
https://github.com/openjs-foundation/cross-project-council/pull/414,
and community input.

Committer: Tobie Langel <tobie@unlockopen.com>