Chore: Add missing logos, convert to svg, add Ajv
A number of project logos were missing, they have been added. In addition, the project logos were linked to a png, which was pixellating at small size. The .svg was available in the artwork repo, so the links have been updated. Finally, Ajv had not been added to the list as an incubating project.

Signed-off-by: Brian Warner <brian@bdwarner.com>