doc: move MylesBorins to emeritus (#719)
Life and work has thrown a bunch of new challenges and expectations my way and I unfortunately don't have the
time to keep up with the commitments required of a regular member. It is pretty amazing that the foundation and
the CPC are in a place where I feel 100% confident in stepping back.

I'll still be around and thanks to our egalitarian governance I'll still be able to participate in all the meetings 🎉

I've moved myself to emeritus, a new section, something we've done over at Node.js that I quite like. I think
it is nice to keep track of former members.

What a wild ride this has been!