country of origin
This is distinct from nationality, ethnicity, and race. For example, someone might be accused of harboring negative loyalties or political inclinations based on the political status of their country of origin, even if they have changed their nationality. It can differ from ethnicity or race in that the attack might be based more on politics rather than culture or racial identity.