chore: update remark tools
The remark tools are very much out of date (and are causing warnings in
the GitHub interface and the command line). This updates them to the
latest versions.

I confirmed that the linting still works by removing the newline at the
end of README.md and seeing that the same warning that it currently
generates is generated with this change.