Update Governance links from HEAD to main (#817)
* Update Governance links from HEAD to main

I am finding that when I click on a link and then want to edit that file, I am viewing it at a commit rather than on the main branch. This makes it so that the file appears uneditable. You simply need to switch to the main branch and then can edit but it may be confusing to folks.

* Update GOVERNANCE.md

Co-authored-by: Eemeli Aro <eemeli@gmail.com>

Co-authored-by: Eemeli Aro <eemeli@gmail.com>