meta: add VoltrexMaster as a regular member
I would like to nominate myself as a regular member.

I've been an active member of the [Node.js organization](https://github.com/nodejs) for about a year now and been helping to maintain Node.js, triaging and helping users/contributors and reviewing to help things land.

I would like to participate and lend a hand to CPC and it's future.