Update project infrastructure documentation
This commit removes references to the Directory generator, which has been deprecated. The LF now
has a centrally maintained tool that does this for us.

There are also a few other changes to the project tooling guidance that have occurred over the past
few years, and the documentation had gotten a little stale. This commit includes updates where
appropriate.

Signed-off-by: Brian Warner <brian@bdwarner.com>