Remove Michael from regular members
I'll not have enough time to be active in the CPC for at least a while so I'll step down for now. 

I was waiting until the new meeting generator tool was in place as the version I was running needed me to have access, but since that is up and ready it's time to step down.

Thanks for the opportunity to contribute over the years and maybe I'll be back a some point.

Signed-off-by: Michael Dawson <mdawson@devrus.com>